manual serialization over memcpy for server struct
Since server structs are going to be bigger than client structs, it
doesnt really make sense to serialize and deserialize them by just
memcpying, since the sizeof the server one is > than the sizeof the
client.  I think this will still work if all of the server specific
fields are at the end of the struct, but it will definitely result in
wasted bandwith (copying stuff we don't care about with memcpy, and
transfering it over the network, just for it to be ignored).