shared:types:hash: consider already set keys
Old behavior:
When a value in the hash is unset, the 'set' flag is turned off.  Now
you go to set a new value with key you just unset.  The first unset slot
discovered will be used, regardless of the old key of that slot.  You
now have one key pointing to two values.

To solve this, keep track of if slots have their keys set, and never
overwrite a key.