don't call opengl render if window not focused
the problem is that, at least on sway, if the window is in a different
workspace then the call to glfwSwapBuffers blocks until the window is
visible again.  GLFW_VISIBLE does not get updated when the window moves
(maybe this is another bug in glfw wayland implementation?).  The only
reason the block is a problem, is because the game loop is all one
thread so networking is blocked too, and thus the connection to the
server dies.